Move over all the migrated tests over to suites
