+++ packages/@glimmer/test-helpers/lib/suites.ts
@@ -1,3 +1,11 @@
 export * from './suites/initial-render';
 export * from './suites/components';
 export * from './suites/ember-components';
+export * from './suites/has-block';
+export * from './suites/debugger';
+export * from './suites/has-block-params-helper';
+export * from './suites/in-element';
+export * from './suites/scopes';
+export * from './suites/shadowing';
+export * from './suites/with-dynamic-vars';
+export * from './suites/yield';
